GP-ChatBot/requirements.txt
flask
flask-cors
openai==0.27.8
gunicorn
python-dotenv
